use x and y scaling factors (thanks to @Guldoman)

+++ src/renderer.h
@@ -29,7 +29,7 @@ float ren_font_group_get_size(RenFont **font);
 void ren_font_group_set_size(RenFont **font, float size);
 void ren_font_group_set_tab_size(RenFont **font, int n);
 float ren_font_group_get_width(RenFont **font, const char *text, size_t len);
-float ren_draw_text(RenFont **font, const char *text, size_t len, float x, int y, RenColor color);
+float ren_draw_text(RenFont **font, const char *text, size_t len, float x, float y, RenColor color);
 
 void ren_draw_rect(RenRect rect, RenColor color);
